+++ include/settings.h
@@ -45,7 +45,7 @@
 // into actual stepper steps that move the stepper motor. It takes 2,181.76 steps to rotate the knob 1 full revolution. with hardware version 1.
 // Incline_Multiplier may be able to be removed in the future by dividing ShiftSteps by ~200 to get this value but we're not quite ready
 // to make that commitment yet.
-#define INCLINE_MULTIPLIER 3.0
+#define INCLINE_MULTIPLIER 6.0
 
 // Minumum value for power correction factor user setting
 #define MIN_PCF .5
@@ -55,7 +55,7 @@
 
 // Default Stepper Power.
 // Stepper peak current in ma. This is hardware restricted to a maximum of 2000ma on the TMC2225. RMS current is less.
-#define DEFAULT_STEPPER_POWER 1500
+#define DEFAULT_STEPPER_POWER 900
 
 // Default Shift Step. THe amount to move the stepper motor for a shift press.
 #define DEFAULT_SHIFT_STEP 1000
@@ -73,7 +73,7 @@
 // I.E. If the difference between ERG target and Current watts were 30, and the Shift step is defined as 600 steps,
 // and ERG_Sensitivity were 1.0, ERG mode would move the stepper motor 600 steps to compensate. With an ERG_Sensitivity of 2.0, the stepper
 // would move 1200 steps to compensate, however ERG_Sensitivity values much different than 1.0 imply shiftStep has been improperly configured.
-#define ERG_SENSITIVITY 1.0
+#define ERG_SENSITIVITY 5.0
 
 // Number of watts per shift expected by ERG mode for it's calculation. The user should target this number by adjusting Shift Step until WATTS_PER_SHIFT
 // is obtained as closely as possible during each shift.
@@ -83,7 +83,7 @@
 #define DEFAULT_MIN_WATTS 50
 
 // Default Max Watts that the brake on the spin bike can absorb from the user.
-#define DEFAULT_MAX_WATTS 700
+#define DEFAULT_MAX_WATTS 800
 
 // Wattage at which to automatically assume minimum brake resistance.
 #define MIN_WATTS 50